refactor: 댓글 작성 관련 api community 폴더로 이동
